# server/src/__tests__/api/swagger.yaml
openapi: 3.0.0
info:
  title: GenKiKoi API
  version: 1.0.0
  description: API documentation for GenKiKoi service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://genkikoi-backend.onrender.com
    description: Production server
...
paths:
...
  /api/doctors:
    get:
      summary: Get all doctors
      tags: [Doctors]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of doctors
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Doctor'
        '404':
          description: No doctors available
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Danh sách bác sĩ trống!
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đã xảy ra lỗi server
    post:
      summary: Create a new doctor
      tags: [Doctors]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - gender
                - email
                - specialization
                - licenseNumber
                - yearOfExperience
                - movingService
              properties:
                fullName:
                  type: string
                  description: Full name of the doctor
                gender:
                  type: string
                  description: Gender of the doctor
                email:
                  type: string
                  description: Email of the doctor
                specialization:
                  type: string
                  description: Specialization of the doctor
                licenseNumber:
                  type: string
                  description: License number of the doctor
                yearOfExperience:
                  type: number
                  description: Years of experience of the doctor
                movingService:
                  type: boolean
                  description: Whether the doctor provides moving service
      responses:
        '201':
          description: Doctor created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Nhân viên được thêm thành công!
                  data:
                    $ref: '#/components/schemas/Doctor'
        '400':
          description: Doctor already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bác sĩ đã tồn tại
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đã xảy ra lỗi server
...
  /api/doctors/{doctorId}:
    get:
      summary: Get a doctor by ID
      tags: [Doctors]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
          description: ID of the doctor
      responses:
        '200':
          description: Doctor information
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Doctor'
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Không tìm thấy bác sĩ
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đã xảy ra lỗi server
  /api/doctor-schedules:
    get:
      summary: Get all doctor schedules
      tags: [Doctors]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of doctor schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DoctorSchedule'
        '404':
          description: No schedules found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Không tìm thấy lịch làm việc nào
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đã xảy ra lỗi khi lấy lịch trình
  /api/doctor-schedules/{doctorId}:
    get:
      summary: Get schedule by doctor ID
      tags: [Doctors]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
          description: ID of the doctor
      responses:
        '200':
          description: Doctor schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DoctorSchedule'
        '404':
          description: Schedule not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Không tìm thấy lịch trình
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đã xảy ra lỗi khi lấy lịch trình
...
  /api/appointments/doctor/{doctorId}:
    get:
      summary: Get appointments by doctor ID
      tags: [Appointments]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
          description: ID of the doctor
      responses:
        '200':
          description: List of appointments for the doctor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        appointmentId:
                          type: string
                        customerName:
                          type: string
                        serviceName:
                          type: string
                        phoneNumber:
                          type: string
                        appointmentDate:
                          type: string
                        status:
                          type: string
                        paymentStatus:
                          type: string
                        isFeedback:
                          type: boolean
        '404':
          description: No appointments found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Không tìm thấy cuộc hẹn
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lỗi khi lấy danh sách cuộc hẹn
...
  /api/feedback/doctor/{doctorId}:
    get:
      summary: Get feedbacks by doctor ID
      tags: [Feedback]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: doctorId
          required: true
          schema:
            type: string
          description: ID of the doctor
      responses:
        '200':
          description: List of feedbacks for the doctor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        customerName:
                          type: string
                        serviceName:
                          type: string
                        feedbackDate:
                          type: string
                        rating:
                          type: number
                        comment:
                          type: string
        '404':
          description: No feedbacks found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...